requirements-local.txt
--index-url="https://artifactory.oci.oraclecorp.com/api/pypi/global-dev-pypi/simple"
--extra-index-url="https://artifactory.oci.oraclecorp.com/api/pypi/ocas-service-platform-dev-pypi-local/simple/"
--trusted-host="artifactory.oci.oraclecorp.com"
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
circuitbreaker==1.4.0
cryptography==42.0.8
generative-ai-service-bmc-python-client==0.2.330
idna==3.10
oci==2.128.0+3068
oml4py==2.10.4
oracledb==3.2.0
pycparser==2.22
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
urllib3==2.5.0



